try fix for whisperx hindi model and remove faster-whisper dependancy

+++ app/service/transcription/speaker_diarization_linux_alt.py
@@ -197,7 +197,7 @@ def process_audio(job_id, bucket, key):
 
 def transcribe_with_whisperx(wav_path, device="cuda"):
     # Load WhisperX model
-    model = whisperx.load_model(WHISPER_MODEL, device, compute_type="float16" if device == "cuda" else "float32")
+    model = whisperx.load_model(WHISPER_MODEL, device, compute_type="float32" if device == "cuda" else "float32")
 
     # Transcribe
     transcription = model.transcribe(
